Change project structure * Move DTOs
