+++ src/main/java/com/wafflestudio/draft/model/request/CreateRoomRequest.java
@@ -1,0 +1,18 @@
+package com.wafflestudio.draft.model.request;
+
+
+import lombok.Data;
+
+import javax.validation.constraints.NotNull;
+import java.time.LocalDateTime;
+
+@Data
+public class CreateRoomRequest {
+    @NotNull
+    private LocalDateTime startTime;
+    @NotNull
+    private LocalDateTime endTime;
+    @NotNull
+    private String name;
+}
+
