+++ src/main/java/com/wafflestudio/draft/model/request/GetRoomsRequest.java
@@ -1,0 +1,15 @@
+package com.wafflestudio.draft.model.request;
+
+import lombok.Data;
+import org.springframework.format.annotation.DateTimeFormat;
+
+import java.time.LocalDateTime;
+
+@Data
+public class GetRoomsRequest {
+    private String name;
+    @DateTimeFormat(pattern = "yyyy-MM-dd HH:mm:ss")
+    private LocalDateTime startTime;
+    @DateTimeFormat(pattern = "yyyy-MM-dd HH:mm:ss")
+    private LocalDateTime endTime;
+}
