+++ src/main/java/com/wafflestudio/draft/model/request/SetDeviceRequest.java
@@ -1,0 +1,13 @@
+package com.wafflestudio.draft.model.request;
+
+import lombok.Data;
+
+import javax.validation.constraints.NotNull;
+
+@Data
+public class SetDeviceRequest {
+    @NotNull
+    private String email;
+    @NotNull
+    private String deviceToken;
+}
