+++ src/main/java/com/wafflestudio/draft/model/request/SetPreferenceRequest.java
@@ -1,0 +1,14 @@
+package com.wafflestudio.draft.model.request;
+
+import com.wafflestudio.draft.model.Preference;
+import lombok.Data;
+
+import java.util.List;
+
+@Data
+public class SetPreferenceRequest {
+    private String regionName;
+    private List<Preference> preferences;
+}
+
+
